fix : file 분류
